Fix data types.
If you're accumulating the lengths of strings, as returned by strlen(),
and passing the result to malloc(), it should be accumulated in a
size_t.

+++ tcpdump.c
@@ -985,7 +985,7 @@ static char *
 copy_argv(char **argv)
 {
 	char **p;
-	u_int len = 0;
+	size_t len = 0;
 	char *buf;
 	char *src, *dst;
